+++ src/todo/TodoModule.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { TaskController } from './controller/TaskController';
+import { TaskService } from './service/TaskService';
 
 @Module({
   imports: [],
   controllers: [TaskController],
-  providers: [],
+  providers: [TaskService],
   exports: [],
 })
 export class TodoModule {}
