+++ src/todo/entity/Task.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, JoinColumn, ManyToOne } from 'typeorm';
 import { TaskStatus } from '../enum/TaskStatus';
-import { TraceableEntity } from './TraceableEntity';
+import { TraceableEntity } from '../../core/entity/TraceableEntity';
 import { TaskList } from './TaskList';
 
 @Entity()
